+++ Package.swift
@@ -7,9 +7,9 @@ let package = Package(
         .library(name: "ReactiveSwift", targets: ["ReactiveSwift"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/antitypical/Result.git", from: "3.2.1"),
-        .package(url: "https://github.com/Quick/Quick.git", from: "1.2.0"),
-        .package(url: "https://github.com/Quick/Nimble.git", from: "7.0.3"),
+        .package(url: "https://github.com/antitypical/Result.git", from: "4.0.0"),
+        .package(url: "https://github.com/Quick/Quick.git", from: "1.3.0"),
+        .package(url: "https://github.com/Quick/Nimble.git", from: "7.1.2"),
     ],
     targets: [
         .target(name: "ReactiveSwift", dependencies: ["Result"], path: "Sources"),