+++ Sources/ValidatingProperty.swift
@@ -177,32 +177,14 @@ public final class ValidatingProperty<Value, ValidationError: Swift.Error>: Muta
 	) {
 		self.init(MutableProperty(initial), with: other, validator)
 	}
-
-	/// Create a `ValidatingProperty` that presents a mutable validating
-	/// view for an inner mutable property.
+	
+	/// Create a `ValidatingProperty` which validates mutations before
+	/// committing them.
 	///
 	/// The proposed value is only committed when `valid` is returned by the
 	/// `validator` closure.
 	///
 	/// - note: `inner` is retained by the created property.
-	///
-	/// - parameters:
-	///   - inner: The inner property which validated values are committed to.
-	///   - other: The property that `validator` depends on.
-	///   - validator: The closure to invoke for any proposed value to `self`.
-	public convenience init<U, E>(
-		_ inner: MutableProperty<Value>,
-		with other: ValidatingProperty<U, E>,
-		_ validator: @escaping (Value, U) -> Decision
-	) {
-		self.init(inner, with: other, validator)
-	}
-
-	/// Create a `ValidatingProperty` that validates mutations before
-	/// committing them.
-	///
-	/// The proposed value is only committed when `valid` is returned by the
-	/// `validator` closure.
 	///
 	/// - parameters:
 	///   - initial: The initial value of the property. It is not required to
@@ -214,10 +196,28 @@ public final class ValidatingProperty<Value, ValidationError: Swift.Error>: Muta
 		with other: ValidatingProperty<U, E>,
 		_ validator: @escaping (Value, U) -> Decision
 	) {
+		self.init(MutableProperty(initial), with: other, validator)
+	}
+
+	/// Create a `ValidatingProperty` that presents a mutable validating
+	/// view for an inner mutable property.
+	///
+	/// The proposed value is only committed when `valid` is returned by the
+	/// `validator` closure.
+	///
+	/// - parameters:
+	///   - inner: The inner property which validated values are committed to.
+	///   - other: The property that `validator` depends on.
+	///   - validator: The closure to invoke for any proposed value to `self`.
+	public convenience init<U, E>(
+		_ inner: MutableProperty<Value>,
+		with other: ValidatingProperty<U, E>,
+		_ validator: @escaping (Value, U) -> Decision
+	) {
 		// Capture only `other.result` but not `other`.
 		let otherValidations = other.result
 
-		self.init(initial) { input in
+		self.init(inner) { input in
 			let otherValue: U
 
 			switch otherValidations.value {