+++ Tests/ReactiveSwiftTests/SignalProducerLiftingSpec.swift
@@ -722,6 +722,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 				observer.send(value: 0)
 				expect(lastValue) == 0
 			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.skip(until: .init(value: ()))
+			}
 		}
 
 		describe("take") {
@@ -1017,6 +1022,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 				expect(completed) == true
 				expect(lastValue).to(beNil())
 			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.take(until: .init(value: ()))
+			}
 		}
 
 		describe("takeUntilReplacement") {
@@ -1076,6 +1086,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 				expect(completed) == false
 				replacementObserver.sendCompleted()
 				expect(completed) == true
+			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.take(untilReplacement: .init(value: 0))
 			}
 		}
 
@@ -1418,6 +1433,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 
 				expect(valueReceived) == "1a"
 			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.sample(with: .init(value: 0))
+			}
 		}
 
 		describe("sampleOn") {
@@ -1525,6 +1545,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 					disposable.dispose()
 					expect(payloadFreed) == true
 				}
+			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.sample(on: .init(value: ()))
 			}
 		}
 
@@ -1658,6 +1683,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 				sampleeObserver.sendInterrupted()
 				expect(event).to(beNil())
 			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.withLatest(from: .init(value: 0))
+			}
 		}
 
 		describe("combineLatestWith") {
@@ -1699,6 +1729,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 
 				otherObserver.sendCompleted()
 				expect(completed) == true
+			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.combineLatest(with: .init(value: 0))
 			}
 		}
 
@@ -1766,6 +1801,11 @@ class SignalProducerLiftingSpec: QuickSpec {
 				rightObserver.send(value: "foo")
 				expect(completed) == true
 				expect(result) == [ "0foo" ]
+			}
+
+			it("should be able to fallback to SignalProducer for contextual lookups") {
+				_ = SignalProducer<Int, NoError>.empty
+					.zip(with: .init(value: 0))
 			}
 		}